Final version

+++ prog1/strLib.h
@@ -1,5 +1,5 @@
-long str2val(char *str){
-    long res = 0 ; 
+long long str2val(char *str){
+    long long  res = 0 ; 
     int len;
     int k;
     
@@ -9,10 +9,11 @@ long str2val(char *str){
         return -1;
     }
     for(k=0;k<len;k++){
-        //        printf("c = %c \t",str[k]);
+       // printf("c = %c %d \t",str[k],str[k] - '0');
         res = res * 10 + str[k] - '0';
+       // printf("res = %lf \n",(double)res);
     }
-     printf("Returning res = %ld for %s\n",res,str);
+     printf("Returning res = %lld for %s\n",res,str);
     return res;
 }
